+++ internal/collector/signal.go
